Added sub-component
