--- src/pages/engagement.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Engagement = () => {
-  return <div>Engagement</div>;
-};
-
-export default Engagement;
