--- src/pages/reglages.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Reglages = () => {
-  return <div>Reglages</div>;
-};
-
-export default Reglages;
